+++ graphbrain/agents/claims.py
@@ -108,7 +108,7 @@ class Claims(Agent):
                     claims = edge.edges_with_argrole('r')
                     if len(subjects) == 1 and len(claims) >= 1:
                         subject = find_concept(subjects[0])
-                        if subject and is_proper_concept(subject):
+                        if subject and has_proper_concept(subject):
                             actor = main_coref(self.hg, subjects[0])
                             if self._is_actor(actor):
                                 for claim in claims: