+++ graphbrain/meaning/actors.py
@@ -2,6 +2,7 @@ from graphbrain.meaning.corefs import main_coref
 
 
 def is_actor(hg, edge):
+    """Checks if the edge is a coreference to an actor."""
     if edge.type()[0] == 'c':
         return hg.exists(('actor/p/.', main_coref(hg, edge)))
     else:
@@ -9,6 +10,7 @@ def is_actor(hg, edge):
 
 
 def find_actors(hg, edge):
+    """Returns set of all coreferences to actors found in the edge."""
     actors = set()
     if is_actor(hg, edge):
         actors.add(main_coref(hg, edge))