+++ graphbrain/meaning/concepts.py
@@ -1,4 +1,6 @@
 def find_concept(edge):
+    """Find the shalowest, left-most concept in the edge. Can be the edge
+    itself."""
     if edge.type()[0] == 'c':
         return edge
     elif not edge.is_atom():
@@ -7,12 +9,13 @@ def find_concept(edge):
         return None
 
 
-def is_proper_concept(edge):
+def has_proper_concept(edge):
+    """Check if the concept either is a proper edge, or contains one."""
     if edge.is_atom():
         return edge.type()[:2] == 'cp'
     else:
         for subedge in edge[1:]:
-            if is_proper_concept(subedge):
+            if has_proper_concept(subedge):
                 return True
         return False